# cross_entropy: modify Liger triton kernel to be cudagraph compatible