--- a/tritonbench/operators/cross_entropy/operator.py
+++ b/tritonbench/operators/cross_entropy/operator.py
@@ -12,7 +12,7 @@ from tritonbench.utils.triton_op import (
 )
 
 try:
-    from liger_kernel.transformers.cross_entropy import LigerCrossEntropyLoss
+    from .liger_cross_entropy import LigerCrossEntropyLoss
 except ModuleNotFoundError:
     LigerCrossEntropyLoss = None