… tritonbench/operators/cross_entropy/liger_cross_entropy.py