Previously, running cross_entropy Liger kernels with Triton do_bench_cudagraph fails with `CUDA error: operation failed due to a previous error during capture`. This PR removes the incompatibility from the Liger kernel to make it cudagraph compatible.

Performance of the Liger kernel stays the same:
On H100:
Before:
```
$ python run.py   --op cross_entropy --metrics speedup,accuracy               --latency-measure-mode profiler               --only liger               --only-match-mode prefix-with-baseline
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 6/6 [00:20<00:00,  3.35s/it]
        (B, T, V)    liger_cross_entropy_loss-speedup    liger_cross_entropy_loss-accuracy
-----------------  ----------------------------------  -----------------------------------
  (8, 2048, 4096)                            0.848115                                    1
  (8, 2048, 8192)                            1.12636                                     1
 (8, 2048, 16384)                            0.909232                                    1
 (8, 2048, 32768)                            1.48511                                     1
 (8, 2048, 65536)                            1.55754                                     1
(8, 2048, 131072)                            1.3279                                      1
          average                            1.20904                                     1
```

After:
```
$ python run.py   --op cross_entropy --metrics speedup,accuracy \
              --latency-measure-mode profiler \
              --only liger \
              --only-match-mode prefix-with-baseline
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 6/6 [00:23<00:00,  3.95s/it]
        (B, T, V)    liger_cross_entropy_loss-speedup    liger_cross_entropy_loss-accuracy
-----------------  ----------------------------------  -----------------------------------
  (8, 2048, 4096)                            0.84033                                     1
  (8, 2048, 8192)                            1.10772                                     1
 (8, 2048, 16384)                            0.903518                                    1
 (8, 2048, 32768)                            1.47731                                     1
 (8, 2048, 65536)                            1.55918                                     1
(8, 2048, 131072)                            1.32759                                     1
          average                            1.20261                                     1
```