@yf225 2025-09-26
> It is good for now, but in longer term it will lose track with upstream since the upstream Liger kernel repo might change.
> 
> What exactly makes the upstream Liger kernel incompatible with CUDA Graph? Is it because of dynamic memory allocation?

These are the incompatibilities I found:
- Upstream still pulls GPU scalars back to Python during the forward path: n_non_ignore = target_mask.sum().item() forces a sync when the graph is being captured, which CUDA Graph prohibits (liger_kernel/ops/cross_entropy.py:288).
- The same host sync happens for the class-weight accumulators: sum_non_ignore_weight = ... .sum().item() and weight_sum = weight.sum().item() both materialize device values on the CPU during capture (liger_kernel/ops/cross_entropy.py:300, liger_kernel/
  ops/cross_entropy.py:301).

I'll follow up to file an issue / a PR to Liger repo.